venue refactor
