+++ docs/docs/auto-docs/components/Venues/Modal/VenueModal/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`__namedParameters`): `Element`
 
-Defined in: [src/components/Venues/VenueModal.tsx:56](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/VenueModal.tsx#L56)
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:56](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L56)
 
 A modal component for creating or updating venue information.
