+++ docs/docs/auto-docs/components/Venues/Modal/VenueModal/interfaces/InterfaceVenueModalProps.md
@@ -1,0 +1,63 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceVenueModalProps
+
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:15](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L15)
+
+## Properties
+
+### edit
+
+> **edit**: `boolean`
+
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:21](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L21)
+
+***
+
+### onHide()
+
+> **onHide**: () => `void`
+
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L17)
+
+#### Returns
+
+`void`
+
+***
+
+### orgId
+
+> **orgId**: `string`
+
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:19](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L19)
+
+***
+
+### refetchVenues()
+
+> **refetchVenues**: () => `void`
+
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:18](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L18)
+
+#### Returns
+
+`void`
+
+***
+
+### show
+
+> **show**: `boolean`
+
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:16](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L16)
+
+***
+
+### venueData?
+
+> `optional` **venueData**: [`InterfaceQueryVenueListItem`](../../../../../utils/interfaces/interfaces/InterfaceQueryVenueListItem.md)
+
+Defined in: [src/components/Venues/Modal/VenueModal.tsx:20](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/Venues/Modal/VenueModal.tsx#L20)
