+++ src/components/Venues/Modal/VenueModal.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { Button, Form, Modal } from 'react-bootstrap';
-import styles from '../../style/app.module.css';
+import styles from '../../../style/app.module.css';
 import { toast } from 'react-toastify';
 import { useTranslation } from 'react-i18next';
 import { useMutation } from '@apollo/client';
